+++ actix-server/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased - 2023-xx-xx
 
+- ServerBuilder: add support for MPTCP (optional).
 - Minimum supported Rust version (MSRV) is now 1.60.
 
 ## 2.2.0 - 2022-12-21